+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.4] - 2021-30-10
+### Changed
+- Moved panel to View catergory
+- Dynamic naming moved to panel itself > Panel Header has font size issue
+
+
 ## [0.0.3] - 2021-29-10
 ### Added
 - Reset operator > set all inputs to default
@@ -23,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.4]:https://github.com/schroef/World_Control/releases/tag/v0.0.4
 [0.0.3]:https://github.com/schroef/World_Control/releases/tag/v0.0.3
 [0.0.2]:https://github.com/schroef/World_Control/releases/tag/v0.0.2
 [0.0.1]:https://github.com/schroef/World_Control/releases/tag/v0.0.1